// MercurionWebNode/src/app_modules/auth/models/interfaces/i-session.interface.ts
import { UUID } from "crypto";
import { AuthProvider } from "src/app_modules/sso/models/enums/auth-provider.enum";
import type { SessionDeviceInfo } from '@mercurion/rest-contracts'

export interface ISession {
    sessionId: UUID
    userId: UUID
    deviceId: string
    createdAt: number
    expiresAt: number
    lastAccessedAt: number
    IP: string
    valid: boolean
    sessionDeviceInfo: SessionDeviceInfo
    fingerprint: string
    location: string
    provider: AuthProvider
}

export type ISSO_SessionActivationData = Pick<ISession, 'IP' | 'fingerprint' | 'sessionDeviceInfo' | 'location' | 'deviceId'>